requirements.txt
Flask==3.0.0
passlib==1.7.4
pandas==2.1.3
py-cpuinfo==9.0.0
python-dotenv==1.0.0
requests==2.31.0
spacy==3.7.2
numpy==1.25.2
pdf2image==1.16.3
azure-ai-formrecognizer==3.3.1
Pillow==10.0.1
